+++ contrib/cirrus/lib.sh
@@ -135,6 +135,7 @@ setup_rootless() {
     req_env_vars GOPATH GOSRC SECRET_ENV_RE
 
     ROOTLESS_USER="${ROOTLESS_USER:-some${RANDOM}dude}"
+    ROOTLESS_UID=""
 
     local rootless_uid
     local rootless_gid
@@ -158,6 +159,7 @@ setup_rootless() {
     cd $GOSRC || exit 1
     # Guarantee independence from specific values
     rootless_uid=$[RANDOM+1000]
+    ROOTLESS_UID=$rootless_uid
     rootless_gid=$[RANDOM+1000]
     msg "creating $rootless_uid:$rootless_gid $ROOTLESS_USER user"
     groupadd -g $rootless_gid $ROOTLESS_USER
