+++ pkg/domain/entities/reports/scp.go
@@ -1,0 +1,5 @@
+package reports
+
+type ScpReport struct {
+	Id string `json:"Id"` //nolint:revive,stylecheck
+}
